requirements.txt
altgraph==0.17
astroid==2.4.2
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
future==0.18.2
html5lib==1.1
idna==2.10
isort==5.5.4
keyboard==0.13.5
lazy-object-proxy==1.4.3
mccabe==0.6.1
MouseInfo==0.1.3
numpy==1.19.2
opencv-python==4.4.0.44
pefile==2019.4.18
Pillow==7.2.0
py3dns==3.2.1
PyAutoGUI==0.9.50
PyGetWindow==0.0.8
pyinstaller==4.0
pyinstaller-hooks-contrib==2020.8
pylint==2.6.0
PyMsgBox==1.0.8
PyMuPDF==1.17.7
PyPDF2==1.26.0
pyperclip==1.8.0
pypiwin32==223
PyRect==0.1.4
PyScreeze==0.1.26
PyTweening==1.0.3
pywin32==228
pywin32-ctypes==0.2.0
reportlab==3.5.53
requests==2.24.0
six==1.15.0
toml==0.10.1
urllib3==1.25.10
validate-email==1.3
webencodings==0.5.1
win10toast==0.9
wrapt==1.12.1
zc.lockfile==2.0
